+++ CHANGELOG.md
@@ -4,6 +4,7 @@ Release Date: _unreleased_
 - Fix invalid `index_in_epoch` and `index_in_round` for genesis TX.
 - Fix virtual identity and account details.
 - Fix how TX total fee paid is calculated.
+- Add `from_state_version` to `validator_emission_statistics`.
 
 ### 0.5.1 RCNet v3 revision 2
 Release Date: 01.09.2023